gitlab

+++ Models/product/ProductImage.py
@@ -9,7 +9,7 @@ from component.snowFlakeId import snowFlack
 class ProductImage(Base):
     __tablename__ = 'product_image'
     __table_args__ = (Index('product_id_order_index', "product_id", "image_order"),)
-    product_id=Column(INTEGER,ForeignKey('variant_static.id'),server_default="0")
+    product_id=Column(BIGINT,ForeignKey('variant_static.id'),server_default="0")
     image_url=Column(VARCHAR(512))
     image_alt=Column(VARCHAR(255))
     image_order=Column(INTEGER,server_default="0")
